+++ server/src/main/java/org/example/server/mapper/RestaurantMapper.java
@@ -29,7 +29,7 @@ public interface RestaurantMapper {
      * @param markerId
      * @return
      */
-    List<Restaurant> queryByMarkerId(int markerId);
+    List<Restaurant> queryByMarkerId(long markerId);
 
 
     /**
